check forget password and update register

+++ app/src/main/java/com/visualinnovate/almursheed/tourist/hire/viewmodel/HireViewModel.kt
@@ -65,7 +65,7 @@ class HireViewModel @Inject constructor(
         viewModelScope.launch {
             safeApiCall {
                 // Make your API call here using Retrofit service or similar
-                apiService.getAllDriversByDistCityId(SharedPreference.getStateId()!!.toInt())
+                apiService.getAllDriversByDistCityId(SharedPreference.getStateId()!!)
             }.collect {
                 when (it) {
                     is ResponseHandler.Success -> {
